requirements.txt
certifi==2021.10.8
cmake==3.22.0
cycler==0.11.0
dlib==19.22.1
fonttools==4.28.3
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
opencv-contrib-python==4.5.4.60
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pyusb==1.2.1
scikit-learn==1.0.1
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
threadpoolctl==3.0.0
tomli==1.2.2
wincertstore==0.2
